+++ src/main/scala/org/broadinstitute/dsde/snoop/dataaccess/DataAccess.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.snoop.data
+package org.broadinstitute.dsde.snoop.dataaccess
 
 import org.broadinstitute.dsde.snoop.util.Reflection
