+++ src/main/scala/org/broadinstitute/dsde/snoop/model/Submission.scala
@@ -1,0 +1,16 @@
+package org.broadinstitute.dsde.snoop.model
+
+import java.sql.Timestamp
+
+/**
+ * Created by dvoet on 3/30/15.
+ */
+case class Submission
+(
+  submissionId: String,
+  submissionDate: Option[Timestamp] = None,
+  modifiedDate: Option[Timestamp] = None,
+  callbackUri: String,
+  status: String,
+  id: Option[Int] = None
+)
